fix: fixed warnings preventing windows build
(cherry picked from commit fb5eea4be17c90039d4da3b7c5bef4133f70142e)

+++ src/modules/dmx/util/artnet_send.cpp
@@ -17,7 +17,7 @@
 namespace caspar {
     namespace dmx {
 
-        void send_dmx_data(int port, const std::wstring& host, int universe, std::vector<std::uint8_t> data) {
+        void send_dmx_data(unsigned short port, const std::wstring& host, int universe, std::vector<std::uint8_t> data) {
             size_t length = data.size();
             if (length > 512 || length % 2 != 0) {
                 // TODO: throw error
