+++ src/modules/dmx/util/artnet_send.h
@@ -16,6 +16,6 @@
 namespace caspar {
     namespace dmx {
 
-        void send_dmx_data(int port, const std::wstring& host, int universe, std::vector<uint8_t> data);
+        void send_dmx_data(unsigned short port, const std::wstring& host, int universe, std::vector<uint8_t> data);
     }
 } // namespace caspar::dmx
